Switch to libargparse and split into files.
