+++ apps/avifgainmaputil/.clang-format
