+++ tests/test_cmd_avifgainmaputil.sh
@@ -49,7 +49,7 @@ pushd ${TMP_DIR}
 
   "${AVIFGAINMAPUTIL}" extractgainmap "${INPUT_AVIF_GAINMAP}" "${AVIF_OUTPUT}" --quality 50
   "${AVIFGAINMAPUTIL}" extractgainmap "${INPUT_AVIF_GAINMAP}" "${JPEG_OUTPUT}"
-  "${AVIFGAINMAPUTIL}" extractgainmap  --speed 9 "${INPUT_AVIF_GAINMAP}" "${PNG_OUTPUT}"
+  "${AVIFGAINMAPUTIL}" extractgainmap --speed 9 "${INPUT_AVIF_GAINMAP}" "${PNG_OUTPUT}"
 
 popd
